fix(m7.14): post-merge followups — 4 code fixes from PR #52 review
1. queue.rs resolve_pending: gate old_suffix match on parse_queue_filename
   session_id absence (prevents spurious new-format match via suffix overlap).
2. config.rs QueueConfig: remove eprintln! spam from effective_pending_ttl()
   and effective_decided_retention() hot-path accessors — clamp silently.
3. queue.rs list_pending_filtered: clarify legacy prune_old_entries comment
   (safety net for S4 lazy GC, not redundant).
4. cli/main.rs queue prune: change older_than_days from i64 to u64 to
   prevent negative-wrap when multiplying by 24.

Also fixes a time-dependent test (backcompat entry with fixed 2026-04-01
timestamp swept by prune_old_entries on May 2026 — use Utc::now()).

+++ colmena-cli/src/main.rs
@@ -147,7 +147,7 @@ enum QueueAction {
     Prune {
         /// Maximum age in days (default: 7)
         #[arg(long, default_value = "7")]
-        older_than: i64,
+        older_than: u64,
     },
 }
 
@@ -1138,11 +1138,11 @@ fn run_queue_list(session_filter: Option<&str>) -> Result<()> {
     Ok(())
 }
 
-fn run_queue_prune(older_than_days: i64) -> Result<()> {
+fn run_queue_prune(older_than_days: u64) -> Result<()> {
     let config_dir = colmena_core::paths::default_config_dir();
     // M7.14: queue prune now cleans decided/ (manual reclaim).
     // Stale pending/ → decided/ is handled automatically by lazy GC + Stop sweep.
-    let retention_hours = (older_than_days * 24) as u64;
+    let retention_hours = older_than_days * 24;
     let now = chrono::Utc::now();
     let deleted = colmena_core::queue::purge_expired_decided(&config_dir, retention_hours, now)?;
     if deleted == 0 {
